improve OOP feature
