+++ src/test/java/seedu/duke/MapTest.java
@@ -1,5 +1,6 @@
 package seedu.duke;
 
+import control.findNearest;
 import control.loadData;
 import org.junit.jupiter.api.Test;
 
@@ -12,7 +13,7 @@ class MapTest {
     void findFacilityByName_validLibrary() throws InvalidCommandException {
         Map map = new Map();
         loadData ld = map.getDataController();
-        String outputFacility = map.findFacilityByName(ld, "library1").getName();
+        String outputFacility = findNearest.findFacilityByName(ld, "library1").getName();
         String expectedFacilityFound = "library1";
         assertEquals(expectedFacilityFound, outputFacility);
 
@@ -22,7 +23,7 @@ class MapTest {
     void findFacilityByName_validCanteen() throws InvalidCommandException {
         Map map = new Map();
         loadData ld = map.getDataController();
-        String outputFacility = map.findFacilityByName(ld, "canteen1").getName();
+        String outputFacility = findNearest.findFacilityByName(ld, "canteen1").getName();
         String expectedFacilityFound = "canteen1";
         assertEquals(expectedFacilityFound, outputFacility);
 
@@ -32,7 +33,7 @@ class MapTest {
     void findFacilityByName_validLT() throws InvalidCommandException {
         Map map = new Map();
         loadData ld = map.getDataController();
-        String outputFacility = map.findFacilityByName(ld, "lectureTheater1").getName();
+        String outputFacility = findNearest.findFacilityByName(ld, "lectureTheater1").getName();
         String expectedFacilityFound = "lectureTheater1";
         assertEquals(expectedFacilityFound, outputFacility);
 
@@ -42,7 +43,7 @@ class MapTest {
         Map map = new Map();
         loadData ld = map.getDataController();
         assertThrows(InvalidCommandException.class, () -> {
-            map.findFacilityByName(ld, "lectureTheater3");
+            findNearest.findFacilityByName(ld, "lectureTheater3");
         });
 
     }
